moved all dependencies to root package.json, deleted client and server package.json, moved all test files into the same file as the file they test
