+++ package.json
@@ -8,13 +8,12 @@
   "license": "ISC",
   "scripts": {
     "build": "webpack",
-    "clean": "rm -r node_modules && cd src/client/ && rm -r node_modules && cd ../server/ && rm -r node_modules && cd ../../",
+    "clean": "rm -r node_modules",
     "dev": "webpack-dev-server",
     "dev:test": "npm-run-all --parallel dev test:watch",
-    "postinstall": "cd src/client/ && npm i && cd ../server/ && npm i & cd ../../",
     "start": "babel-watch src/server",
-    "test": "mocha --compilers js:babel-core/register",
-    "test:watch": "mocha --watch --compilers js:babel-core/register"
+    "test": "mocha --reporter progress buildScripts/testSetup.js \"src/**/*.test.js\"",
+    "test:watch": "npm run test -- --watch"
   },
   "devDependencies": {
     "babel-core": "^6.26.3",
@@ -22,6 +21,7 @@
     "babel-preset-env": "^1.7.0",
     "babel-preset-react": "^6.24.1",
     "babel-preset-stage-0": "^6.24.1",
+    "babel-register": "^6.26.0",
     "babel-watch": "^2.0.7",
     "chai": "^4.1.2",
     "chalk": "^2.4.1",
@@ -31,10 +31,8 @@
     "eslint-plugin-mocha": "^4.12.1",
     "eslint-plugin-react": "^7.11.1",
     "html-webpack-plugin": "^2.30.1",
-    "mocha": "^4.1.0",
+    "mocha": "^5.2.0",
     "npm-run-all": "^4.1.3",
-    "react": "^16.4.2",
-    "react-dom": "^16.4.2",
     "react-hot-loader": "^3.1.3",
     "style-loader": "^0.19.1",
     "webpack": "^3.12.0",
@@ -42,13 +40,25 @@
   },
   "dependencies": {
     "array-flatten": "^2.1.1",
+    "async": "^2.6.1",
     "babel-cli": "^6.26.0",
     "classnames": "2.2.6",
     "dateformat": "^3.0.3",
+    "express": "^4.16.3",
     "lodash": "^4.17.10",
+    "material-ui": "^0.20.2",
     "numeral": "^2.0.6",
+    "prop-types": "^15.6.2",
+    "react": "^16.4.2",
     "react-addons-css-transition-group": "15.6.2",
+    "react-dom": "^16.4.2",
+    "react-fontawesome": "^1.6.1",
+    "react-redux": "^5.0.7",
+    "react-router-dom": "^4.3.1",
     "react-spinners": "^0.2.6",
+    "recharts": "^1.1.0",
+    "redux": "^3.7.2",
+    "redux-thunk": "^2.3.0",
     "request": "^2.88.0",
     "yahoo-finance": "^0.3.5"
   }
