--- src/client/package.json
@@ -1,17 +1,0 @@
-{
-  "name": "client",
-  "dependencies": {
-    "lodash": "^4.17.10",
-    "material-ui": "^0.20.2",
-    "numeral": "^2.0.6",
-    "prop-types": "^15.6.2",
-    "react": "^16.4.2",
-    "react-dom": "^16.4.2",
-    "react-fontawesome": "^1.6.1",
-    "react-redux": "^5.0.7",
-    "react-router-dom": "^4.3.1",
-    "recharts": "^1.1.0",
-    "redux": "^3.7.2",
-    "redux-thunk": "^2.3.0"
-  }
-}
