--- test/index.js
@@ -1,15 +1,0 @@
-import {
-  formatValueFromStateAndPropsTest
-} from './utils/formatting/stringFormatting';
-import {
-  formatDateForMaxStockDataTest
-} from './utils/dateUtils/dateFormatting';
-import {
-  addCommasTest,
-  padSingleDigitWithZeroTest
-} from './utils/formatting/numberFormatting';
-
-formatValueFromStateAndPropsTest();
-padSingleDigitWithZeroTest();
-formatDateForMaxStockDataTest();
-addCommasTest();
